requirements.txt
asarPy==1.0.1
asgiref==3.9.1
Cython==3.1.3
Django==5.2.6
django-filter==25.1
djangorestframework==3.16.1
Faker==37.6.0
Mako==1.3.10.dev0
Markdown==3.8.2
MarkupSafe==3.0.2
meson==1.5.2
packaging==25.0
pycairo==1.28.0
Pygments==2.19.2
PyGObject==3.50.1
python-dotenv==1.1.1
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
validate_docbr==1.11.1
wheel==0.46.1
